Hide endpoints in dedicated file
