+++ src/Backend/Tomorrowify/Configuration/Configuration.cs
@@ -1,6 +1,6 @@
 ﻿namespace Tomorrowify.Configuration;
 
-public sealed class Configuration
+public sealed class TomorrowifyConfiguration
 {
     public string? ClientSecret { get; set; }
 }
